Add a button switch in the hallway as a rfxtrx binary sensor
The iPad in the hallway has been replaced with an outlet and a button
switch. Use this button to turn on different scenes.

Disable auto add because it may cause some problems I have read.

Also send fire an event even if the state is the same. This makes it
possible to push the button several times.