Replace crystal lamp with a Philips Hue light with colors
The guy at Hornbach double checked the price when he saw that the light bulb
cost 485 SEK. I had a hard time convincing him that it I had come to the
conclusion that it was worth it. Now that I see the light bulb in the hallway,
I am very happy with it.
The old Z-wave controller from Nexa was not very nice to me. I am so happy
that I did not buy a new one but instead bought this!