Add Halloween automation
This was very successful on this Halloween. When a visitor came into the
garden all the lights in the garden and the lights in the house visible
from the garden were switched off.

Short after a sound of thunder started playing in an outdoor Heos One
Speaker and my Twinkly lights started to blink.

Some 30 seconds a few lead lights went on again and after 2-3 minutes
all the lights went on again and everything was prepared for the next
visitor.

The candy was stored in a big chest in the garden, very near a grave
yard for animals. A few ghosts, pumpkins and other lights were also part
of this installation.